pyproject.toml
[project]
name = "brecup"
version = "0.0.5"
authors = [{ name = "aquanjsw", email = "zhdlcc@gmail.com" }]
description = "Simplify the BililiveRecorder → biliup-rs workflow."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["pyyaml", "loguru"]

[project.scripts]
brecup = "brecup:main.main"

[project.urls]
homepage = "https://github.com/aquanjsw/brecup"
